Generate data sets under least favorable parameter configurations

Generates a (simulation) instance, a list of multiple datasets to be processed (analyzed) with process_instance. Ground truth parameters (Sensitvity & Specificity) are least-favorable in the sense that the type-I error rate of the subsequently applied multiple test procedures is maximized.

This function is only needed for simulation via batchtools, not relevant in interactive use!

Usage

generate_instance_lfc(
  nrep = 10,
  n = 100,
  prev = 0.5,
  random = FALSE,
  m = 10,
  se = 0.8,
  sp = 0.8,
  L = 1,
  rhose = 0,
  rhosp = 0,
  cortype = "equi",
  ...,
  data = NULL,
  job = NULL
)

Arguments

nrep

(numeric)
integer, number of instances

n

(numeric)
integer, total sample size

prev

(numeric)
disease prevalence

random

(logical)
fixed prevalence (FALSE) or simple random sampling (TRUE)

m

(numeric)
integer, number of candidates

se

(numeric)
sensitivity

sp

(numeric)
specificity

L

(numeric)
worst alternative is computed under side condition Acc <= L

rhose

(numeric)
correlation parameter for sensitivity

rhosp

(numeric)
correlation parameter for specificity

cortype

(character)
correlation type ("equi" or "ak1")

...

(any)
further (named) arguments

data

(NULL)
ignored (for batchtools compatibility)

job

(NULL)
ignored (for batchtools compatibility)

Value

(list)
a single (LFC) simulation instance of length nrep

Details

Utilizes same arguments as draw_data_lfc unless mentioned otherwise above.