Analyze simulated synthetic datasets.

Process data instances, a list of multiple datasets generated via generate_instance_lfc or generate_instance_roc. This function applies evaluate to all datasets.

This function is only needed for simulation via batchtools, not relevant in interactive use!

Usage

process_instance(
  instance = NULL,
  contrast = "cases::define_contrast('raw', NA)",
  benchmark = 0.5,
  alpha = 0.05,
  alternative = "greater",
  adjustment = "none",
  transformation = "none",
  analysis = "co-primary",
  regu = "c(1,1/2,1/4)",
  pars = "list()",
  ...,
  data = NULL,
  job = list(id = NA)
)

Arguments

instance

(list)
generated via generate_instance_lfc or generate_instance_roc.

contrast

(cases_contrast)
specified via define_contrast

benchmark

(numeric)
value to compare against (RHS), should have same length as data or length one if all benchmark values are identical.

alpha

(numeric)
significance level (default: 0.05)

alternative

(character)
specify alternative hypothesis

adjustment

(character)
specify type of statistical adjustment taken to address multiplicity

transformation

(character)
define transformation to ensure results (e.g. point estimates, confidence limits) lie in unit interval ("none" (default) or "logit")

analysis

(character)
"co-primary" (default; only option currently)

regu

(numeric | logical)
vector of length 3, specify type of shrinkage. Alternatively, logical of length one (TRUE := c(2, 1, 1/2), FALSE := c(0, 0, 0))

pars

(list)
further parameters given as named list

...

(any)
additional named parameters

data

(NULL)
ignored (for batchtools compatibility)

job

(NULL)
for batchtools compatibility, do not change

Value

(list)
standardized evaluation results

Details

Utilizes same arguments as evaluate unless mentioned otherwise above.